refactor(indexer): merge ExternalSourceProcessor into LeaderLogProcessor
Prep for re-enabling trie garbage collection (Phase 2 of #5501).
Trie GC on the leader needs a mutex to serialise `TriesDeleted` publishes against `TriesAdded` / `ResolvedTx` / `BlockBoundary` on the replica log.
ESP already had that mutex because `indexTx` (driven from the external-source subscription coroutine) and `processRecords` (driven from the source-log subscription) both publish to the same transactional Kafka producer.
LLP didn't — its single source-log driver had no second writer to race against.

Giving LLP its own mutex was the obvious move, but it would have duplicated the mutex-owning machinery — and, by extension, the `appendToReplica` path, the `finishBlock` bookkeeping, and soon the trie-GC plumbing.
After the earlier `TxIndexer` refactor landed, the only genuinely variant surface between the two processors was: LLP decoded tx ops from the source log and drove the Clojure indexer; ESP ran an `extJob` coroutine driving `ExternalSource.onPartitionAssigned`.
Everything else was unconditionally shared, and the `Tx` / `LegacyTx` / `Attach` / `Detach` branches ESP defensively errored on were inert in practice — ext-source DBs' source logs don't contain those messages; the submit path blocks them.

So the two processors collapse into one, driven either by source-log alone or by source-log + external source.
The mutex lives on the merged class.
A follow-up commit adds trie GC on top with the mutex already in place.

`TxIndexer` turns from a concrete class (the old `TxIndexerImpl`, deleted here) into an interface, and the merged leader processor implements it.
External-source plugins — `modules/postgres-source` — still receive a `TxIndexer` via `ExternalSource.onPartitionAssigned(…, txIndexer)` and call `indexTx` on it; keeping `TxIndexer` as a type means that plugin-facing API stays stable.
`indexTx`'s `txId` and `systemTime` defaults move from instance-state expressions to `null` defaults (interfaces can't express the former), and the impl resolves nulls to `liveIndex.latestCompletedTx` / `instantSource.instant()` — callsites unaffected.
`TxCommitter` is gone: the merged processor owns both `indexTx` and the private `commit` it delegates to.

`QueryOpts` also moves from `TxIndexer` to `OpenTx`.
It's caller-side options for `OpenTx.openQuery` / `OpenTx.executeDml` — not an indexer concern — and keeping it on `TxIndexer` would have forced the Clojure indexer to import the interface just for a data class.

`partition: Int` stays non-null.
Was briefly made `Int?` during the merge to pair with nullable `extSource`, then reverted — threading `partition != null iff extSource != null` through init invariants added noise for no payoff.
Unused `partition = 0` in source-log-only mode is cheap and clear.

The merged class takes back the `LeaderLogProcessor` name — "ExternalSource" no longer describes what it is, only one of its inputs.
Test-class names (`LeaderLogProcessorTest`, `ExternalSourceTest`) are preserved; one covers the merged leader processor, the other covers the external-source-driven path through it.

Out of scope: the trie-GC work itself (`ReplicaMessage.TriesDeleted`, follower replay, per-cycle delete budget, defaulting trie GC on).
Tracked on #5501.

Refs #5501
