+++ core/src/main/kotlin/xtdb/indexer/LeaderLogProcessor.kt
@@ -1,50 +1,75 @@
 package xtdb.indexer
 
+import io.micrometer.core.instrument.Counter
+import kotlinx.coroutines.CancellationException
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.sync.Mutex
+import kotlinx.coroutines.sync.withLock
 import org.apache.arrow.memory.BufferAllocator
 import xtdb.NodeBase
-import xtdb.database.ExternalSourceToken
 import xtdb.api.TransactionKey
 import xtdb.api.TransactionResult
 import xtdb.api.log.*
 import xtdb.api.log.Log.AtomicProducer.Companion.withTx
 import xtdb.api.log.ReplicaMessage.BlockBoundary
+import xtdb.api.log.ReplicaMessage.TriesAdded
 import xtdb.api.storage.Storage
 import xtdb.arrow.Relation
 import xtdb.arrow.asChannel
 import xtdb.database.Database
 import xtdb.database.DatabaseState
 import xtdb.database.DatabaseStorage
+import xtdb.database.ExternalSource
+import xtdb.database.ExternalSourceToken
 import xtdb.error.Anomaly
 import xtdb.error.Interrupted
+import xtdb.indexer.Indexer.Companion.addTxRow
 import xtdb.indexer.TxIndexer.TxResult
 import xtdb.table.TableRef
+import xtdb.time.InstantUtil.asMicros
+import xtdb.time.InstantUtil.fromMicros
 import xtdb.tx.deserializeUserMetadata
-import xtdb.util.*
 import xtdb.util.StringUtil.asLexDec
 import xtdb.util.StringUtil.asLexHex
+import xtdb.util.asPath
+import xtdb.util.debug
+import xtdb.util.error
+import xtdb.util.logger
+import xtdb.util.trace
+import xtdb.util.warn
 import java.nio.ByteBuffer
 import java.time.Duration
+import java.time.Instant
+import java.time.InstantSource
 import java.time.ZoneId
 import java.time.ZonedDateTime
+import kotlin.coroutines.CoroutineContext
 
 private val LOG = LeaderLogProcessor::class.logger
 
 class LeaderLogProcessor(
     allocator: BufferAllocator,
-    nodeBase: NodeBase,
-    dbStorage: DatabaseStorage,
-    private val replicaProducer: Log.AtomicProducer<ReplicaMessage>,
-    private val dbState: DatabaseState,
+    private val nodeBase: NodeBase,
+    private val dbStorage: DatabaseStorage,
     indexer: Indexer,
     crashLogger: CrashLogger,
+    private val dbState: DatabaseState,
+    private val blockUploader: BlockUploader,
     private val watchers: Watchers,
+    private val extSource: ExternalSource?,
+    private val replicaProducer: Log.AtomicProducer<ReplicaMessage>,
     private val skipTxs: Set<MessageId>,
     private val dbCatalog: Database.Catalog?,
-    private val blockUploader: BlockUploader,
+    partition: Int,
     afterSourceMsgId: MessageId,
     afterReplicaMsgId: MessageId,
+    afterToken: ExternalSourceToken?,
+    private val instantSource: InstantSource = InstantSource.system(),
     flushTimeout: Duration = Duration.ofMinutes(5),
-) : LogProcessor.LeaderProcessor, TxCommitter {
+    ctx: CoroutineContext = Dispatchers.Default,
+) : LogProcessor.LeaderProcessor, TxIndexer {
 
     init {
         require((dbCatalog != null) == (dbState.name == "xtdb")) {
@@ -64,14 +89,8 @@ class LeaderLogProcessor(
 
     private val tracer = nodeBase.tracer?.takeIf { nodeBase.config.tracer.transactionTracing }
 
-    private val txIndexer = TxIndexer(this.allocator, nodeBase, dbStorage, dbState, watchers, committer = this, tracer = tracer)
-
     private val indexer: Indexer.ForDatabase =
-        indexer.openForDatabase(this.allocator, dbState, liveIndex, crashLogger, txIndexer)
-
-    override suspend fun commit(openTx: OpenTx, result: TxResult) {
-        throw UnsupportedOperationException("LeaderLogProcessor.commit — internal-indexer migration not yet complete")
-    }
+        indexer.openForDatabase(this.allocator, dbState, liveIndex, crashLogger, this)
 
     override var pendingBlock: PendingBlock? = null
         private set
@@ -84,6 +103,82 @@ class LeaderLogProcessor(
 
     private val blockFlusher = BlockFlusher(flushTimeout, blockCatalog)
 
+    // Serialises the two input streams — source-log records (via [processRecords]) and external-source txs
+    // (via [indexTx] driven from the subscription coroutine below).
+    // The transactional Kafka producer (replicaProducer) is not safe for concurrent `withTx`, so both paths
+    // must hold this lock before touching it.
+    private val mutex = Mutex()
+
+    private val txErrorCounter: Counter? =
+        nodeBase.meterRegistry?.let { Counter.builder("tx.error").register(it) }
+
+    private val extJob = extSource?.let { source ->
+        CoroutineScope(ctx).launch {
+            try {
+                source.onPartitionAssigned(partition, afterToken, this@LeaderLogProcessor)
+            } catch (e: CancellationException) {
+                throw e
+            } catch (e: Throwable) {
+                watchers.notifyError(e)
+            }
+        }
+    }
+
+    private fun smoothSystemTime(systemTime: Instant): Instant {
+        val lct = liveIndex.latestCompletedTx?.systemTime ?: return systemTime
+        val floor = fromMicros(lct.asMicros + 1)
+        return if (systemTime.isBefore(floor)) floor else systemTime
+    }
+
+    private fun openTx(txKey: TransactionKey, externalSourceToken: ExternalSourceToken?) =
+        OpenTx(allocator, nodeBase, dbStorage, dbState, txKey, externalSourceToken, tracer)
+
+    override fun startTx(txKey: TransactionKey): OpenTx = openTx(txKey, null)
+
+    override suspend fun indexTx(
+        externalSourceToken: ExternalSourceToken?,
+        txId: Long?,
+        systemTime: Instant?,
+        writer: suspend (OpenTx) -> TxResult,
+    ): TxResult {
+        val resolvedTxId = txId ?: ((liveIndex.latestCompletedTx?.txId ?: -1) + 1)
+        val resolvedSystemTime = systemTime ?: instantSource.instant()
+        val txKey = TransactionKey(resolvedTxId, smoothSystemTime(resolvedSystemTime))
+
+        try {
+            val openTx = openTx(txKey, externalSourceToken)
+            val result = try {
+                writer(openTx)
+            } catch (e: Throwable) {
+                openTx.close()
+                throw e
+            }
+
+            when (result) {
+                is TxResult.Committed -> openTx.use {
+                    it.addTxRow(dbName, txKey, null, result.userMetadata)
+                    commit(it, result)
+                }
+
+                is TxResult.Aborted -> {
+                    txErrorCounter?.increment()
+                    openTx.close()
+                    openTx(txKey, externalSourceToken).use { abortTx ->
+                        abortTx.addTxRow(dbName, txKey, result.error, result.userMetadata)
+                        commit(abortTx, result)
+                    }
+                }
+            }
+
+            return result
+        } catch (e: CancellationException) {
+            throw e
+        } catch (e: Throwable) {
+            watchers.notifyError(e)
+            throw e
+        }
+    }
+
     private suspend fun maybeFlushBlock() {
         if (blockFlusher.checkBlockTimeout(blockCatalog, liveIndex)) {
             val flushMessage = SourceMessage.FlushBlock(blockCatalog.currentBlockIndex ?: -1)
@@ -94,6 +189,19 @@ class LeaderLogProcessor(
     private suspend fun appendToReplica(message: ReplicaMessage): Log.MessageMetadata =
         replicaProducer.withTx { tx -> tx.appendMessage(message) }.await()
             .also { latestReplicaMsgId = it.msgId }
+
+    private suspend fun finishBlock(latestProcessedMsgId: MessageId, externalSourceToken: ExternalSourceToken?) {
+        val boundaryMsg =
+            BlockBoundary((blockCatalog.currentBlockIndex ?: -1) + 1, latestProcessedMsgId, externalSourceToken)
+
+        val boundaryMsgId = appendToReplica(boundaryMsg).msgId
+        LOG.debug("[$dbName] block boundary b${boundaryMsg.blockIndex.asLexHex}: source=$latestProcessedMsgId, replica=$boundaryMsgId")
+
+        pendingBlock = PendingBlock(boundaryMsgId, boundaryMsg)
+
+        latestReplicaMsgId = blockUploader.uploadBlock(replicaProducer, boundaryMsgId, boundaryMsg)
+        pendingBlock = null
+    }
 
     private fun resolveTx(
         msgId: MessageId, record: Log.Record<SourceMessage>, msg: SourceMessage
@@ -170,17 +278,6 @@ class LeaderLogProcessor(
         watchers.notifyTx(result, resolvedTx.txId, resolvedTx.externalSourceToken)
     }
 
-    private suspend fun finishBlock(latestProcessedMsgId: MessageId, externalSourceToken: ExternalSourceToken?) {
-        val boundaryMsg =
-            BlockBoundary((blockCatalog.currentBlockIndex ?: -1) + 1, latestProcessedMsgId, externalSourceToken)
-        val boundaryMsgId = appendToReplica(boundaryMsg).msgId
-        LOG.debug("[$dbName] block boundary b${boundaryMsg.blockIndex.asLexHex}: source=$latestProcessedMsgId, replica=$boundaryMsgId")
-        pendingBlock = PendingBlock(boundaryMsgId, boundaryMsg)
-
-        latestReplicaMsgId = blockUploader.uploadBlock(replicaProducer, boundaryMsgId, boundaryMsg)
-        pendingBlock = null
-    }
-
     private suspend fun handleResolvedTx(resolvedTx: ReplicaMessage.ResolvedTx) {
         val txId = resolvedTx.txId
 
@@ -192,122 +289,160 @@ class LeaderLogProcessor(
             finishBlock(txId, resolvedTx.externalSourceToken)
     }
 
+    private suspend fun commit(openTx: OpenTx, result: TxResult) {
+        val txKey = openTx.txKey
+        val externalSourceToken = openTx.externalSourceToken
+        val tableData = openTx.serializeTableData()
+
+        mutex.withLock {
+            liveIndex.commitTx(openTx)
+
+            val resolvedTx = ReplicaMessage.ResolvedTx(
+                txKey.txId, txKey.systemTime,
+                committed = when (result) {
+                    is TxResult.Committed -> true
+                    is TxResult.Aborted -> false
+                },
+                error = (result as? TxResult.Aborted)?.error,
+                tableData, dbOp = null,
+                externalSourceToken = externalSourceToken,
+            )
+
+            appendToReplica(resolvedTx)
+
+            val txResult = when (result) {
+                is TxResult.Committed -> TransactionResult.Committed(txKey)
+                is TxResult.Aborted -> TransactionResult.Aborted(txKey, result.error)
+            }
+            watchers.notifyTx(txResult, latestSourceMsgId, resolvedTx.externalSourceToken)
+
+            if (liveIndex.isFull())
+                finishBlock(latestSourceMsgId, resolvedTx.externalSourceToken)
+        }
+    }
+
     override suspend fun processRecords(records: List<Log.Record<SourceMessage>>) {
         maybeFlushBlock()
 
         for (record in records) {
-            val msgId = record.msgId
-            LOG.trace { "[$dbName] leader: message $msgId (${record.message::class.simpleName})" }
+            mutex.withLock {
+                val msgId = record.msgId
+                LOG.trace { "[$dbName] leader: message $msgId (${record.message::class.simpleName})" }
 
-            try {
-                when (val msg = record.message) {
-                    is SourceMessage.Tx -> {
-                        val resolved = resolveTx(msgId, record, msg)
-                        handleResolvedTx(msg.externalSourceToken?.let { resolved.copy(externalSourceToken = it) }
-                            ?: resolved)
-                    }
-
-                    is SourceMessage.LegacyTx -> handleResolvedTx(resolveTx(msgId, record, msg))
-
-                    is SourceMessage.FlushBlock -> {
-                        val expectedBlockIdx = msg.expectedBlockIdx
-                        if (expectedBlockIdx != null && expectedBlockIdx == (blockCatalog.currentBlockIndex ?: -1L)) {
-                            finishBlock(msgId, watchers.externalSourceToken)
+                try {
+                    when (val msg = record.message) {
+                        is SourceMessage.Tx -> {
+                            val resolved = resolveTx(msgId, record, msg)
+                            handleResolvedTx(msg.externalSourceToken?.let { resolved.copy(externalSourceToken = it) }
+                                ?: resolved)
                         }
-                        latestSourceMsgId = msgId
-                        watchers.notifyMsg(msgId)
-                    }
 
-                    is SourceMessage.AttachDatabase -> {
-                        val txKey = TransactionKey(msgId, record.logTimestamp)
-                        val error = if (dbCatalog != null) {
-                            try {
-                                dbCatalog.attach(msg.dbName, msg.config)
-                                null
-                            } catch (e: Anomaly.Caller) {
-                                LOG.debug(e) { "[$dbName] leader: attach database '${msg.dbName}' failed at $msgId" }
-                                e
+                        is SourceMessage.LegacyTx -> handleResolvedTx(resolveTx(msgId, record, msg))
+
+                        is SourceMessage.FlushBlock -> {
+                            val expectedBlockIdx = msg.expectedBlockIdx
+                            if (expectedBlockIdx != null && expectedBlockIdx == (blockCatalog.currentBlockIndex ?: -1L)) {
+                                finishBlock(msgId, watchers.externalSourceToken)
                             }
-                        } else null
+                            latestSourceMsgId = msgId
+                            watchers.notifyMsg(msgId)
+                        }
 
-                        val resolvedTx = indexer.addTxRow(txKey, error)
-                            .let { if (error == null) it.copy(dbOp = DbOp.Attach(msg.dbName, msg.config)) else it }
+                        is SourceMessage.AttachDatabase -> {
+                            val txKey = TransactionKey(msgId, record.logTimestamp)
+                            val error = if (dbCatalog != null) {
+                                try {
+                                    dbCatalog.attach(msg.dbName, msg.config)
+                                    null
+                                } catch (e: Anomaly.Caller) {
+                                    LOG.debug(e) { "[$dbName] leader: attach database '${msg.dbName}' failed at $msgId" }
+                                    e
+                                }
+                            } else null
 
-                        appendToReplica(resolvedTx)
+                            val resolvedTx = indexer.addTxRow(txKey, error)
+                                .let { if (error == null) it.copy(dbOp = DbOp.Attach(msg.dbName, msg.config)) else it }
 
-                        val result =
-                            if (error == null) TransactionResult.Committed(txKey)
+                            appendToReplica(resolvedTx)
+
+                            val result =
+                                if (error == null) TransactionResult.Committed(txKey)
+                                else TransactionResult.Aborted(txKey, error)
+                            latestSourceMsgId = msgId
+                            watchers.notifyTx(result, msgId, null)
+                        }
+
+                        is SourceMessage.DetachDatabase -> {
+                            val txKey = TransactionKey(msgId, record.logTimestamp)
+                            val error = if (dbCatalog != null) {
+                                try {
+                                    dbCatalog.detach(msg.dbName)
+                                    null
+                                } catch (e: Anomaly.Caller) {
+                                    LOG.debug(e) { "[$dbName] leader: detach database '${msg.dbName}' failed at $msgId" }
+                                    e
+                                }
+                            } else null
+
+                            val resolvedTx = indexer.addTxRow(txKey, error)
+                                .let { if (error == null) it.copy(dbOp = DbOp.Detach(msg.dbName)) else it }
+
+                            appendToReplica(resolvedTx)
+
+                            val result = if (error == null) TransactionResult.Committed(txKey)
                             else TransactionResult.Aborted(txKey, error)
-                        latestSourceMsgId = msgId
-                        watchers.notifyTx(result, msgId, null)
-                    }
-
-                    is SourceMessage.DetachDatabase -> {
-                        val txKey = TransactionKey(msgId, record.logTimestamp)
-                        val error = if (dbCatalog != null) {
-                            try {
-                                dbCatalog.detach(msg.dbName)
-                                null
-                            } catch (e: Anomaly.Caller) {
-                                LOG.debug(e) { "[$dbName] leader: detach database '${msg.dbName}' failed at $msgId" }
-                                e
-                            }
-                        } else null
-
-                        val resolvedTx = indexer.addTxRow(txKey, error)
-                            .let { if (error == null) it.copy(dbOp = DbOp.Detach(msg.dbName)) else it }
-
-                        appendToReplica(resolvedTx)
-
-                        val result = if (error == null) TransactionResult.Committed(txKey)
-                        else TransactionResult.Aborted(txKey, error)
-                        latestSourceMsgId = msgId
-                        watchers.notifyTx(result, msgId, null)
-                    }
-
-                    is SourceMessage.TriesAdded -> {
-                        if (msg.storageVersion == Storage.VERSION && msg.storageEpoch == bufferPool.epoch) {
-                            msg.tries.groupBy { it.tableName }.forEach { (tableName, tries) ->
-                                trieCatalog.addTries(
-                                    TableRef.parse(dbState.name, tableName),
-                                    tries,
-                                    record.logTimestamp
-                                )
-                            }
+                            latestSourceMsgId = msgId
+                            watchers.notifyTx(result, msgId, null)
                         }
 
-                        appendToReplica(
-                            ReplicaMessage.TriesAdded(
-                                msg.storageVersion, msg.storageEpoch, msg.tries, sourceMsgId = msgId
+                        is SourceMessage.TriesAdded -> {
+                            if (msg.storageVersion == Storage.VERSION && msg.storageEpoch == bufferPool.epoch) {
+                                msg.tries.groupBy { it.tableName }.forEach { (tableName, tries) ->
+                                    trieCatalog.addTries(
+                                        TableRef.parse(dbState.name, tableName),
+                                        tries,
+                                        record.logTimestamp
+                                    )
+                                }
+                            }
+
+                            appendToReplica(
+                                TriesAdded(
+                                    msg.storageVersion, msg.storageEpoch, msg.tries,
+                                    sourceMsgId = msgId
+                                )
                             )
-                        )
 
-                        latestSourceMsgId = msgId
-                        watchers.notifyMsg(msgId)
-                    }
+                            latestSourceMsgId = msgId
+                            watchers.notifyMsg(msgId)
+                        }
 
-                    // TODO this one's going before release
-                    is SourceMessage.BlockUploaded -> {
-                        latestSourceMsgId = msgId
-                        watchers.notifyMsg(msgId)
+                        // TODO this one's going before release
+                        is SourceMessage.BlockUploaded -> {
+                            latestSourceMsgId = msgId
+                            watchers.notifyMsg(msgId)
+                        }
                     }
+                } catch (e: InterruptedException) {
+                    throw e
+                } catch (e: Interrupted) {
+                    throw e
+                } catch (e: Throwable) {
+                    LOG.error(
+                        e,
+                        "[$dbName] leader: failed to process log record with msgId $msgId (${record.message::class.simpleName})"
+                    )
+                    watchers.notifyError(e)
+                    throw e
                 }
-            } catch (e: InterruptedException) {
-                throw e
-            } catch (e: Interrupted) {
-                throw e
-            } catch (e: Throwable) {
-                LOG.error(
-                    e,
-                    "[$dbName] leader: failed to process log record with msgId $msgId (${record.message::class.simpleName})"
-                )
-                watchers.notifyError(e)
-                throw e
             }
         }
     }
 
     override fun close() {
+        // HACK: we cancel without joining because a blocking join deadlocks under runTest's virtual time.
+        extJob?.cancel()
+        extSource?.close()
         indexer.close()
         allocator.close()
     }
